moved the mongodb config beans from ApplicationConfig to MongoConfig
