+++ src/main/java/uk/gov/companieshouse/api/accounts/ApplicationConfiguration.java
@@ -1,32 +1,16 @@
 package uk.gov.companieshouse.api.accounts;
 
-import com.mongodb.MongoClientOptions;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
 
 /**
- * General application configuration .
+ * General application configuration.
+ *
+ * 1. Defines the application configurations
+ * 2. Enables AspectJ proxy for Spring AOP.
  */
 @Configuration
 @EnableAspectJAutoProxy
 public class ApplicationConfiguration {
 
-    @Autowired
-    private MongoDbConnectionPoolProperties configuration;
-
-    /**
-     * Create a {@link MongoClientOptions} .
-     *
-     * @return A {@link MongoClientOptions} .
-     */
-    @Bean
-    public MongoClientOptions mongoClientOptions() {
-        return MongoClientOptions.builder()
-                .minConnectionsPerHost(configuration.getMinSize())
-                .maxConnectionIdleTime(configuration.getMaxConnectionIdleTimeMS())
-                .maxConnectionLifeTime(configuration.getMaxConnectionLifeTimeMS())
-                .build();
-    }
 }
